requirements.txt
Flask
gunicorn
pandas
numpy
scikit-learn
typing
pyYAML
# -e .